+++ primary-backend/src/main/java/com/blinkflow/primary_backend/dto/FlowResponseDTO.java
@@ -15,7 +15,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor
 public class FlowResponseDTO {
 	@NotNull
-	private int flowID;
+	private Long flowID;
 	@NotBlank
 	private String flowTriggerName;
 	@NotEmpty
