+++ primary-backend/src/main/java/com/blinkflow/primary_backend/repository/FlowActionRepository.java
@@ -5,6 +5,6 @@ import org.springframework.stereotype.Repository;
 import com.blinkflow.primary_backend.model.FlowAction;
 
 @Repository
-public interface FlowActionRepository extends JpaRepository<FlowAction, Integer> {
+public interface FlowActionRepository extends JpaRepository<FlowAction, Long> {
 
 }
